Move configuration into it's own file
